Delete file

+++ app/Http/Controllers/ProductoController.php
@@ -153,13 +153,6 @@ class ProductoController extends Controller
         return redirect()->route('producto.index');
     }
 
-    
-    public function pdf()
-    {
-        $productos = Producto::paginate();
-
-        return view('producto', compact('productos'));
-    }
 
     /**
      * Remove the specified resource from storage.
@@ -171,11 +164,12 @@ class ProductoController extends Controller
     {
         $archivo = Archivo::where('producto_id', $producto->id)->first();
 
-        $path = $archivo->hash;
-
-        Storage::disk('public')->delete($path);
-
-        $archivo->delete();
+        if ($archivo !== null){
+            $path = $archivo->hash;
+            $archivo->delete();
+            Storage::disk('public')->delete($path);
+        }
+        
         $producto->delete();
         notyf()
             ->position('x', 'center')
